+++ cogs/librarian.py
@@ -13,6 +13,8 @@ from core import rubbercog
 class Librarian(rubbercog.Rubbercog):
     """Knowledge and information based commands"""
 
+    # TODO Move czech strings to text.default.json
+
     def __init__(self, bot):
         super().__init__(bot)
 
@@ -37,7 +39,7 @@ class Librarian(rubbercog.Rubbercog):
     @commands.command(aliases=["tyden", "týden", "tyzden", "týždeň"])
     async def week(self, ctx: commands.Context):
         """See if the current week is odd or even"""
-        starting_week = config.starting_week
+        starting_week = config.get("librarian", "starting week")
         cal_week = date.today().isocalendar()[1]
         stud_week = cal_week - starting_week
         even, odd = "sudý", "lichý"
@@ -52,7 +54,7 @@ class Librarian(rubbercog.Rubbercog):
 
     @commands.command(aliases=["pocasi", "pocasie"])
     async def weather(self, ctx, *args):
-        token = config.weather_token
+        token = config.get("librarian", "weather token")
         city = "Brno"
         if len(args) != 0:
             city = " ".join(map(str, args))